+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Added
 
+-   [xy] component `ScatterChargedLabels` to display labels linked to data
+    points
 -   [xy] prop `variant` for `ScatterLabel` to place labels next to data points
     identified according to euclidean distance in xy dimensions, or euclidean
     distance along only one of the axes